[feat] SlackMessageResult에 팩토리메서드 추가
